moved devonfwIdeSelector file to shared
